# education-swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Education App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api/v1
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /sphere:
    # binds a127 app logic to a route
    x-swagger-router-controller: get_competence
    get:
      operationId: getSphere
      description: "Provide list of sphere, filtered out by brand or nationwide. The list support paging"
      # used as the method name of the controller
      parameters:
        - name: name
          in: query
          type: string
        - name: level
          in: query
          type: string
          
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/sphere"
        # responses may fall through to errors
        500:
          description: An unexpected error occured.
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  sphere:
    type: object
    properties:
      name:
        type: string
        description: Short name of sphere
      level:
        type: string
        description: level
...
  Error:
    required:
      - message
    properties:
      message:
        type: string
...